# public/openapi.yaml
openapi: 3.1.0
info:
  title: DBackup API
  version: 3.4.0
  description: |
    REST API for DBackup - a self-hosted database backup automation platform with encryption, compression, and smart retention.

    ## Authentication

    DBackup supports two authentication methods:

    - **Session Authentication** - Used automatically when logged in via the web UI. Session cookies are sent with each request.
    - **API Key Authentication** - For scripts, CI/CD pipelines, and external integrations. Create an API key under **Access Management → API Keys**.

    > API keys do not inherit SuperAdmin privileges. Only explicitly assigned permissions are available.

    ## Rate Limits

    | Request Type | Default Limit |
    |---|---|
    | Authentication (login, etc.) | 5/min per IP |
    | GET requests | 100/min per IP |
    | POST / PUT / DELETE | 20/min per IP |

    Rate limits are configurable in Settings → Rate Limits.
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
...
servers:
  - url: "{protocol}://{host}/api"
    description: DBackup Instance
    variables:
      protocol:
        default: https
        enum: [http, https]
      host:
        default: your-dbackup-instance.com
security:
  - BearerAuth: []
  - SessionAuth: []
...
paths:
...
  /storage/{id}/files:
    get:
      tags: [Storage Explorer]
      summary: List Backup Files
      description: Lists all backup files in a storage destination with enriched metadata.
      operationId: listBackupFiles
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
        - name: typeFilter
          in: query
          schema:
            type: string
          description: Filter by file type
      responses:
        "200":
          description: Backup file list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Storage Explorer]
      summary: Delete Backup File
      operationId: deleteBackupFile
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
                  description: File path to delete
                  example: "backups/mydb_2025-01-15.sql.gz.enc"
      responses:
        "200":
          description: File deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /storage/{id}/restore:
    post:
      tags: [Storage Explorer]
      summary: Restore Backup
      description: |
        Start an asynchronous restore process. Poll the returned `executionId` to track progress.

        - `privilegedAuth` provides elevated credentials for `CREATE DATABASE` permissions.
        - A version guard prevents restoring newer dump formats to older database servers.
        - `databaseMapping` allows renaming databases during multi-DB restores.
      operationId: restoreBackup
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestoreRequest"
      responses:
        "202":
          description: Restore started
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  executionId:
                    type: string
                  message:
                    type: string
                    example: "Restore started"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /storage/{id}/restore-files:
    post:
      tags: [Storage Explorer]
      summary: Restore or Download Archive Contents
      description: |
        Reads files and database dumps out of a seekable archive, fetching only the selected entries from the destination where it supports byte ranges.

        With `target.kind` set to `download` the selection is streamed back. Exactly one entry in `databases` and no `selections` returns that database dump on its own, decrypted and decompressed. Anything else returns a `.tar.gz`, and sending neither `selections` nor `databases` returns the complete snapshot. With `prepare: true` the response is a token instead of the bytes, to be fetched with `GET` on the same path.

        The other targets write files back into a storage destination. Database dumps can only be downloaded.

        **Required permission:** `storage:download` for a download, `storage:restore` otherwise
      operationId: restoreArchiveFiles
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [file, target]
              properties:
                file:
                  type: string
                  description: Backup file path
                  example: "backups/shared-mysql-01/nightly_2026-09-16.tar"
                selections:
                  type: array
                  description: Files to include, per directory source. An entry without paths means that whole source.
                  items:
                    type: object
                    required: [src]
                    properties:
                      src:
                        type: string
                        description: Directory source id from the archive index
                      paths:
                        type: array
                        items:
                          type: string
                databases:
                  type: array
                  description: Database dumps to include, by name. Download targets only.
                  items:
                    type: string
                  example: ["customer_shop_prod"]
                target:
                  type: object
                  required: [kind]
                  properties:
                    kind:
                      type: string
                      enum: [download, origin, storage]
                    configId:
                      type: string
                      description: Storage destination, for kind `storage`
                    basePath:
                      type: string
                      description: Path within that destination, for kind `storage`
                excludePatterns:
                  type: array
                  items:
                    type: string
                dryRun:
                  type: boolean
                  description: Resolve the selection and report its size without reading anything
                prepare:
                  type: boolean
                  description: Return a token for a `GET` download instead of the bytes
                profileIdOverride:
                  type: string
                  description: Vault profile to open the backup with
      responses:
        "200":
          description: The selection as a single dump or a `.tar.gz`, or JSON for `dryRun`, `prepare` and storage targets
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
            application/gzip:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                      fileName:
                        type: string
                        example: "nightly_2026-09-16_customer_shop_prod.sql"
                      fileCount:
                        type: integer
                      databaseCount:
                        type: integer
                      totalBytes:
                        type: integer
                      fullDownload:
                        type: boolean
                      output:
                        type: string
                        enum: [dump, tar]
        "400":
          description: Invalid request, or `databases` combined with a storage target
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    get:
      tags: [Storage Explorer]
      summary: Collect Prepared Download
      description: |
        Streams a download prepared with `prepare: true`. The token belongs to the session or API key that created it and expires after 5 minutes.

        **Required permission:** `storage:download`
      operationId: collectArchiveDownload
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The prepared dump or `.tar.gz`
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "410":
          description: Token expired or unknown
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        API Key authentication. Create an API key under **Access Management → API Keys**.

        Format: `dbackup_xxxxxxxx`
    SessionAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: Session cookie (browser login)
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource ID
  responses:
    Unauthorized:
      description: Missing, invalid, disabled, or expired credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Valid credentials but insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Human-readable error message"
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
...
    FileInfo:
      type: object
      properties:
        name:
          type: string
          example: "mydb_2025-01-15.sql.gz.enc"
        path:
          type: string
        size:
          type: integer
          description: Size in bytes
        lastModified:
          type: string
          format: date-time
        locked:
          type: boolean
          description: Whether the file is currently being written
    RestoreRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          description: Backup file path
          example: "backups/mydb_2025-01-15.sql.gz.enc"
        scope:
          type: string
          enum: [all, databases, files]
          default: all
          description: >-
            Which half of a combined (database + directory) archive to restore. Omitting
            `databaseMapping` or `directoryMapping` means "restore all of them", so a
            restore that should leave one half untouched has to say so here.
        targetSourceId:
          type: string
          description: >-
            Target database adapter ID. Required whenever the restore includes at least one
            database - so not for a directory-only backup, nor at `scope: files`.
        targetDatabaseName:
          type: string
          description: Override target DB name (single-DB backups)
        databaseMapping:
          description: |
            Which databases to restore and under which names. Either an object of renames, `{ "old_name": "new_name" }`, which restores exactly the databases it names, or a list of entries where only entries with `selected: true` are restored. Omit it to restore every database in the backup.
          oneOf:
            - type: object
              additionalProperties:
                type: string
            - type: array
              items:
                type: object
                required: [originalName]
                properties:
                  originalName:
                    type: string
                  targetName:
                    type: string
                    description: Defaults to originalName
                  selected:
                    type: boolean
          example:
            old_db: new_db
        privilegedAuth:
          type: object
          properties:
            user:
              type: string
            password:
              type: string
          description: Elevated credentials for CREATE DATABASE permissions

    # ─── Other ───────────────────────────────────────────────────
...